requirements.txt
cashier==1.3
Django==2.2.2
django-auditlog==0.4.5
django-crispy-forms==1.7.2
django-jsonfield==1.2.0
django-widget-tweaks==1.4.5
mysqlclient==1.4.2.post1
Pillow==6.0.0
PyMySQL==0.9.3
python-dateutil==2.6.0
pytz==2019.1
six==1.12.0
somewhere==0.0.1
sqlparse==0.3.0
